Create README.md
The main purpose of this project is to examine and analyze credit card approvals with various variables including demographics, years of employment, and so on. Different machine learning algorithms such as artificial neural network, xgboost, random forest, support vector machine, and regression models such as logistic regression, LASSO and ridge regression are developed to predict the credit card approvals, while the research questions were generated to help in understanding the data.